requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
mlflow>=2.3.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
PyYAML>=5.4.1
pytest>=6.2.5
httpx>=0.18.2
python-multipart>=0.0.5

-e .